Revert "Basic culling of monitors"
This reverts commit 4d3cce467ed58a1230ec6897f500250140a0f492.

This would be really nice to have, but is sadly much trickier to
implement than I initially anticipated. Minecraft doesn't provide access
to the _actual_ camera, and given the odd things we need to take into
account (third person, "hurt" camera roll, etc...) I'm not going to
bother right now.

I'll look into this again at some point in the future - I'm 90% sure
it's possible, and if not on 1.12, then probably on 1.1{4,5}. Goodness
knows :).