Add a monitor renderer using TBOs
This uses the system described in #409 (or at least, how I understand
it), to render monitors in a more efficient manner.

Each monitor is backed by a texture buffer object (TBO) which contains
a relatively compact encoding of the terminal state. This is then
rendered using a shader, which consumes the TBO and uses it to index
into main font texture.

My OpenGL skills are pretty much nonexistent, so the implementation of
this is no doubt terrible. However, the performance so far is
outstanding compared with the current VBO renderer, as it transmits
significantly less data to the GPU.