Always reset to the default shader
We're somewhat broken with optifine (monitors render transparently), so
there's no point trying to restore the previous shader.

Ideally the BEST renderer would use VBOs when a shader pack is enabled,
and use the default TBOs otherwise. However, it's not entirely clear how
to do that*, so this'll do for now.

I guess we could check if the current shader is non-0, and fall back if
so, but it feels bodgey. People will just have to change their config
for now.