Added Javadoc for currently undocumented functions
This PR adds some documentation for APIs that did not have docs in the
source yet. This includes:
* the drive peripheral
* the FS API
* the OS PAI
* the printer peripheral
* the speaker peripheral

The term API also needs documenting, but the TermAPI.java file does not
have the actual functions inside it, so they'll need to go elsewhere.
The turtle API also needs some further documentation, but I don't know
much about it. The official CC wiki can fill in the holes eventually.