Fix getItemDetail showing raw translation keys.
The dedicated server always uses the default Language instance, as it's
setter method is stripped from the server jar. So if we mixin to this
instance's creation and load it with mods' en_us lang files, we will be
able to "translate" text on the server. Translate in quotes as we're
only loading en_us files, but mojang was the one who left out
translations on the dedicated server we shouldn't feel too bad.