Fix finding inventories below y=0.
Thankfully, the surface level of redstone ready flat worlds is now below
y=0 or, I wouldn't have noticed.

I reformatted some things to keep the diffs down against CC-Tweaked.
It might be nicer to make this return an ItemStorage as most
places we grab the Container and immediately wrap it.