Correctly instantiate blocks with their BE type
Previously new blocks were being passed their BlockEntityType directly
when constructed. However, due to the circular dependency between block
entities and blocks, the BlocKEntityType was null.

Normally this'd mean that no BE was created when the block is placed.
However, Block.newBlockEntity was overridden on all subclasses to ensure
that the correct BE was created, /except/ for command computers, which
just returned a normal computer TE.

This changes all the block constructrs to accept a
Supplier<BlockEntityType<?>>, and always uses this to create the block
entity.

Fixes #23.