Fix two bugs with monitors on dedicated server (#45).
- ClientMonitors were being created on the server. This caused a crash
when TileMonitors unload and attempt to clean up their client side
buffers because the method to do that only exists on the client. We
don't have the split semantics of load and handleUpdateTag that forge
has, so our TileMonitor#load method has to do double duty and check if
the level is client side before doing client side stuff.
- Monitor contents were never sent to clients connected to a dedicated
server because MonitorWatcher was never initialized on the dedicated
server! (My bad...) To fix, I moved its initialization to the common
setup.